requirements.txt
crewai==0.177.0
crewai-tools==0.0.1
langchain-openai
tavily-python
python-dotenv
pyyaml
